requirements.txt
streamlit
python-docx
langchain-google-vertexai